move parse_query to utils.py
